cartoon/requirements.txt
absl-py==2.1.0
addict==2.4.0
aiofiles==23.2.1
albumentations==1.3.1
altair==5.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
basicsr==1.4.2
beautifulsoup4==4.12.3
black==24.8.0
bleach==6.1.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.2.1
customtkinter==5.2.0
cycler==0.12.1
Cython==3.0.11
darkdetect==0.8.0
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@5b72c27ae39f99db75d43f18fd1312e1ea934e60
diffusers==0.30.0
easydict==1.13
exceptiongroup==1.2.2
executing==2.0.1
facexlib==0.3.0
fastapi==0.112.1
fastjsonschema==2.20.0
ffmpy==0.4.0
filelock==3.15.4
filterpy==1.4.5
flatbuffers==24.3.25
fonttools==4.53.1
fqdn==1.5.1
fsspec==2024.6.1
future==1.0.0
fvcore==0.1.5.post20221221
gast==0.4.0
gdown==5.2.0
gfpgan==1.3.8
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.34.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gradio==4.41.0
gradio_client==1.3.0
grpcio==1.65.5
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.6
humanfriendly==10.0
hydra-core==1.3.2
idna==3.7
imageio==2.35.1
importlib_metadata==8.2.0
importlib_resources==6.4.3
insightface==0.7.3
iopath==0.1.9
ipykernel==6.29.5
ipython==8.26.0
isoduration==20.11.0
jax==0.4.31
jaxlib==0.4.31
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keras==2.13.1
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.43.0
lmdb==1.5.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.14
mistune==3.0.2
ml-dtypes==0.4.0
mpmath==1.3.0
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.4.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
notebook_shim==0.2.4
numba==0.60.0
numpy==1.24.3
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.14.0
onnxruntime-silicon==1.13.1
opencv-contrib-python==4.10.0.84
opencv-python==4.8.0.74
opencv-python-headless==4.10.0.84
opennsfw2==0.10.2
opt-einsum==3.3.0
optree==0.12.1
orjson==3.10.7
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
Pillow==10.0.0
platformdirs==4.2.2
portalocker==2.10.1
prettytable==3.11.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==4.23.4
psutil==5.9.5
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocotools==2.0.8
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.1.1
qudida==0.0.4
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
ruff==0.6.1
safetensors==0.4.4
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
selenium==4.23.1
semantic-version==2.10.0
Send2Trash==1.8.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.0
soupsieve==2.6
stack-data==0.6.3
starlette==0.38.2
streamlit==1.37.1
sympy==1.13.2
tabulate==0.9.0
tb-nightly==2.18.0a20240818
tenacity==8.5.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.13.0
tensorflow-metal==1.1.0
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2024.8.10
tinycss2==1.3.0
tk==0.1.0
tkinterdnd2-universal==1.7.3
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.4.1
tqdm==4.65.0
traitlets==5.14.3
transformers==4.44.0
trio==0.26.2
trio-websocket==0.11.1
typer==0.12.4
types-python-dateutil==2.9.0.20240316
typing_extensions==4.5.0
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.2
uvicorn==0.30.6
watchdog==4.0.2
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
wrapt==1.16.0
wsproto==1.2.0
yacs==0.1.8
yapf==0.40.2
zipp==3.20.0
